// generated/api/engine/models/renderedComponent.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * BDUI API
 * Спецификация OpenAPI для Backend-Driven UI  приложения

 * OpenAPI spec version: 1.0.0
 */
import type { RenderedRow } from './renderedRow';
import type { RenderedBox } from './renderedBox';
import type { RenderedColumn } from './renderedColumn';
import type { RenderedText } from './renderedText';
import type { RenderedInput } from './renderedInput';
import type { RenderedImage } from './renderedImage';
import type { RenderedSpacer } from './renderedSpacer';
import type { RenderedProgressBar } from './renderedProgressBar';
import type { RenderedSwitch } from './renderedSwitch';
import type { RenderedButton } from './renderedButton';

/**
 * Полиморфный UI-компонент
 */
export type RenderedComponent =
  | (RenderedRow & {
      type: string;
    })
  | (RenderedBox & {
      type: string;
    })
  | (RenderedColumn & {
      type: string;
    })
  | (RenderedText & {
      type: string;
    })
  | (RenderedInput & {
      type: string;
    })
  | (RenderedImage & {
      type: string;
    })
  | (RenderedSpacer & {
      type: string;
    })
  | (RenderedProgressBar & {
      type: string;
    })
  | (RenderedSwitch & {
      type: string;
    })
  | (RenderedButton & {
      type: string;
    });